backend/requirement.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
av==13.1.0
beautifulsoup4==4.14.3
cachetools==6.2.2
cartesia==2.0.17
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloup==3.0.8
colorama==0.4.6
decorator==5.2.1
docstring_parser==0.17.0
fastapi==0.123.5
ffmpeg-python==0.2.0
frozenlist==1.8.0
future==1.0.0
glcontext==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.129.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.6.0
google-cloud-texttospeech==2.33.0
google-crc32c==1.7.1
google-genai==1.53.0
google-generativeai==0.8.5
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
idna==3.11
isosurfaces==0.1.2
iterators==0.2.0
manim==0.19.1
ManimPango==0.6.1
mapbox_earcut==2.0.0
markdown-it-py==4.0.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
multidict==6.7.0
networkx==3.6
numpy==2.3.5
packaging==25.0
pillow==12.0.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyglet==2.1.11
pyglm==2.8.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scipy==1.16.3
screeninfo==0.8.1
shapely==2.1.2
six==1.17.0
skia-pathops==0.9.0
smallestai==4.1.0
soupsieve==2.8
srt==3.5.3
starlette==0.50.0
svgelements==1.9.6
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
fastapi
uvicorn
motor
svix
python-dotenv
pydantic
qdrant-client
httpx
PyMuPDF
python-multipart
pdfplumber